overload predict() to allow for different input types
I (just like the example of the original library) want to input float values into the tflite model.
The previous version to my knowledge only takes int8. So the quantization would need to happen before inputting the value (which e.g. does not happen in the example).
idk if this is a proper way of solving the issue, it is just a suggestion and a fix for me to be able to use the library...